Moved load_data script and added 'hydrocron_load' as a module script. Updated tests to load data from path relative to test file. Minor formatting updates. Added a few missing dependencies.
