+++ .flake8
@@ -1,0 +1,3 @@
+[flake8]
+#ignore = ...
+max-line-length=180
