+++ .pylintrc
@@ -60,6 +60,10 @@ property-classes=abc.abstractproperty
 # Naming style matching correct variable names.
 variable-naming-style=snake_case
 
+[FORMAT]
+
+max-line-length=180
+
 [EXCEPTIONS]
 
 # Exceptions that will emit a warning when being caught. Defaults to
