+++ pyproject.toml
@@ -9,13 +9,16 @@ packages = [{include = "hydrocron_db"}]
 
 [tool.poetry.dependencies]
 python = "^3.10"
-boto3 = "^1.26.147"
-botocore = "^1.29.147"
-geopandas = "^0.13.1"
-
+geopandas = "^0.13.2"
+earthaccess = "^0.5.3"
+boto3 = "^1.28"
 
 [tool.poetry.group.test.dependencies]
 pytest = "^7.3.1"
+pytest-dynamodb = "^2.2.3"
+
+[tool.poetry.scripts]
+hydrocron_load = 'hydrocron_db.load_data:main'
 
 [build-system]
 requires = ["poetry-core"]
