+++ tests/test_io_swot_reach_node_shp.py
@@ -6,12 +6,15 @@ Test unpacking a swot reach or node shapefile.
 
 Unit tests for unpacking swot reach and node shapefiles.
 """
+import os.path
+
 from hydrocron_db.io import swot_reach_node_shp
 
-TEST_SHAPEFILE_PATH = (
-    "tests/data/"
-    "SWOT_L2_HR_RiverSP_Reach_548_011_NA_"
-    "20230610T193337_20230610T193344_PIA1_01.zip")
+TEST_SHAPEFILE_PATH = os.path.join(
+    os.path.dirname(os.path.realpath(__file__)),
+    'data',
+    'SWOT_L2_HR_RiverSP_Reach_548_011_NA_20230610T193337_20230610T193344_PIA1_01.zip'
+)
 
 TEST_FILENAME = (
     "SWOT_L2_HR_RiverSP_Reach_548_011_NA_"
